Add details view

+++ domain/src/main/java/com/edreams/android/workshops/kotlin/domain/model/VenueModel.kt
@@ -1,3 +1,3 @@
 package com.edreams.android.workshops.kotlin.domain.model
 
-class VenueModel(val id : String, val name: String, val rating: Double, val photo: String)
+class VenueModel(val id : String, val name: String, val rating: Float, val photo: String)
